[DIFF] Support generating inline comments, and submitting them.
Unfortunately you cannot submit in-line comments using the conduit api.
To we use the API to prepare all the in-line comments...

At this point they would appear as unsubmitted in the web-interface.

We then use some curl foo to submit them, which uses the session cookie
stored in ~/.config/phabrik/cookie.txt